+++ src/app/services/viewport.js
@@ -1,11 +1,9 @@
-'use strict';
+import {fromEvent} from 'rxjs';
+import {startWith} from 'rxjs/operators';
+import {dispatch} from 'iblokz-state';
+import {obj} from 'iblokz-data';
 
-const {fromEvent} = require('rxjs');
-const {startWith} = require('rxjs/operators');
-const {dispatch} = require('iblokz-state');
-const {obj} = require('iblokz-data');
-
-const screenSize = width => width >= 1200
+export const screenSize = width => width >= 1200
 	? 'xl'
 	: width >= 992
 		? 'lg'
@@ -15,22 +13,22 @@ const screenSize = width => width >= 1200
 				? 'sm'
 				: 'xs';
 
-const patchScreen = () => dispatch(state => obj.patch(state, 'viewport.screen', {
+export const patchScreen = () => dispatch(state => obj.patch(state, 'viewport.screen', {
 	width: window.innerWidth,
 	height: window.innerHeight,
 	size: screenSize(window.innerWidth)
 }));
 
-let stop = () => {};
+export let stop = () => {};
 
-const start = () => {
+export const start = () => {
 	const sub = fromEvent(window, 'resize')
 		.pipe(startWith(null))
 		.subscribe(patchScreen);
 	stop = () => sub.unsubscribe();
 };
 
-module.exports = {
+export default {
 	start,
 	stop
 };