+++ src/app/state/layout/index.js
@@ -1,0 +1,20 @@
+export const initial = {
+  toggles: {
+    leftSideBar: true,
+    rightSideBar: false,
+    bottomPanel: false,
+    preview: false,
+    previewConsole: false
+  },
+  dim: {
+    leftSideBar: 260,
+    rightSideBar: 320,
+    bottomPanel: 320,
+    preview: 0.5,
+    previewConsole: 240
+  }
+};
+
+export default {
+  initial
+}