+++ src/app/ui/empty-editor.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const {section, p} = require('iblokz-snabbdom-helpers');
-
-module.exports = () => section('.empty-editor', [
-	p(['Select a file from the sidebar to open it.'])
-]);